First prototype of calculator model checking (work in progress)
