+++ package.json
@@ -76,5 +76,8 @@
     "glob": "^10.3.3",
     "mocha": "^10.2.0",
     "typescript": "^5.1.6"
+  },
+  "dependencies": {
+    "systeminformation": "^5.21.17"
   }
 }
